# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/chats": {
      "get": {
        "operationId": "ChatController_getAllChats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all chats the user participates in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatListItemDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all chats for the current user",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/my-chats-ids": {
      "get": {
        "operationId": "ChatController_getMyChatsIds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of chat IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get IDs of all chats the user participates in",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/start": {
      "post": {
        "operationId": "ChatController_startChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartChatResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - cannot chat with yourself"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a direct message chat with another user",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}/messages": {
      "get": {
        "operationId": "ChatController_getMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prevCursor",
            "required": false,
            "in": "query",
            "description": "Cursor for loading newer messages",
            "schema": {
              "example": "msg-id",
              "type": "string"
            }
          },
          {
            "name": "nextCursor",
            "required": false,
            "in": "query",
            "description": "Cursor for loading older messages",
            "schema": {
              "example": "msg-id",
              "type": "string"
            }
          },
          {
            "name": "aroundMessageId",
            "required": false,
            "in": "query",
            "description": "Message ID to load messages around (for pinned or search)",
            "schema": {
              "example": "msg-id",
              "type": "string"
            }
          },
          {
            "name": "aroundDate",
            "required": false,
            "in": "query",
            "description": "Timestamp to load messages around",
            "schema": {
              "example": "12345678",
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of messages to load (default: 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMessagesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not a member of this chat"
          },
          "404": {
            "description": "Chat or message not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get messages for a specific chat with pagination support",
        "tags": [
          "chats"
        ]
      },
      "post": {
        "operationId": "ChatController_sendMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not a member of this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message to a chat",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}": {
      "get": {
        "operationId": "ChatController_viewChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat retrieved successfully. Returns FullChatDto if user is participant, PreviewChatDto otherwise",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FullChatDto"
                    },
                    {
                      "$ref": "#/components/schemas/PreviewChatDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - cannot view private chats of others"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "View a specific chat (full or preview)",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}/messages/{messageId}": {
      "delete": {
        "operationId": "ChatController_deleteMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "description": "The ID of the message to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not a member of this chat or not the sender"
          },
          "404": {
            "description": "Chat or message not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a message (only sender can delete)",
        "tags": [
          "chats"
        ]
      },
      "patch": {
        "operationId": "ChatController_editMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "description": "The ID of the message to edit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message edited successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not a member of this chat or not the sender"
          },
          "404": {
            "description": "Chat or message not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a message (only sender can edit)",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}/pinned": {
      "get": {
        "operationId": "ChatController_getPinnedMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": {
              "example": "cursor-string",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of pinned messages to load (default: 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pinned messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPinnedMessagesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not a member of this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get pinned messages in a chat",
        "tags": [
          "chats"
        ]
      }
    },
...
    "/chats/create-group": {
      "post": {
        "operationId": "ChatController_createGroupChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group chat created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGroupChatResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - group must have at least one other member"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new group chat",
        "tags": [
          "chats"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat Backend API",
    "description": "API for chat application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "ChatListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Chat ID"
          },
          "title": {
            "type": "string",
            "description": "Chat title"
          },
          "isGroup": {
            "type": "boolean",
            "description": "Whether this is a group chat"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat creation date"
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last message update time"
          },
          "messages": {
            "description": "Latest message in the chat",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "participants": {
            "description": "Chat participants",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "isGroup",
          "createdAt",
          "participants"
        ]
      },
      "StartChatDto": {
        "type": "object",
        "properties": {
          "otherUserId": {
            "type": "string",
            "description": "User ID of the person to start a chat with",
            "example": "ckx9q8l9o0001z0f9g8h3abcd"
          }
        },
        "required": [
          "otherUserId"
        ]
      },
...
    }
  }
}